fix: Fix the repo selected is incorrect after search
