+++ src/stores/devops/scm.js
@@ -143,19 +143,16 @@ export default class SCMStore extends BaseStore {
       get(this.orgList.data[_activeRepoIndex], 'repositories.nextPage') || 1
 
     const organizationName =
-      this.orgList.data[_activeRepoIndex].name ||
-      this.orgList.data[_activeRepoIndex].key
+      scmType === 'bitbucket-server'
+        ? this.orgList.data[_activeRepoIndex].key
+        : this.orgList.data[_activeRepoIndex].name
 
     this.getRepoListLoading = true
 
     const result = await this.request.get(
       `${this.getBaseUrlV2({
         cluster,
-      })}scms/${scmType}/organizations/${
-        scmType === 'bitbucket-server'
-          ? `~${organizationName}`
-          : organizationName
-      }/repositories/?${getQueryString(
+      })}scms/${scmType}/organizations/${organizationName}/repositories/?${getQueryString(
         {
           ...this.orgParams,
           pageNumber,
